+++ packages/messaging-ui/jest.setup.ts
@@ -1,2 +1,3 @@
-import '@testing-library/jest-dom/extend-expect';
+import '@testing-library/jest-dom/extend-expect'
+import './src/mocks/matchMedia.mock' // required by react-slick
 export default {}
