+++ packages/messaging-ui/package.json
@@ -5,6 +5,7 @@
   "publisher": "botpress",
   "main": "dist/index.js",
   "module": "dist/index.esm.js",
+  "types": "./src/typings.d.ts",
   "dependencies": {
     "html-truncate": "^1.2.2",
     "mime": "^2.5.2",
@@ -39,6 +40,7 @@
     "@types/react": "^17.0.22",
     "@types/react-intl": "^2.3.18",
     "@types/react-select": "^4.0.17",
+    "@types/react-slick": "^0.23.5",
     "babel-jest": "^27.2.1",
     "jest": "^27.2.1",
     "jest-transform-css": "^2.1.0",
